creamos una db relacional, primero valida y crea

+++ backend/app/models/mappings.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import BigInteger, Column, Integer, String
 
 from app.database import Base
 
@@ -7,7 +7,7 @@ class Channel(Base):
     __tablename__ = "channels"
 
     id = Column(Integer, primary_key=True, index=True)
-    channel_id = Column(Integer, unique=True, nullable=False, index=True)
+    channel_id = Column(BigInteger, unique=True, nullable=False, index=True)
     name = Column(String(100), nullable=False)
 
     def __repr__(self):
@@ -18,7 +18,7 @@ class Agent(Base):
     __tablename__ = "agents"
 
     id = Column(Integer, primary_key=True, index=True)
-    agent_id = Column(Integer, unique=True, nullable=False, index=True)
+    agent_id = Column(BigInteger, unique=True, nullable=False, index=True)
     name = Column(String(100), nullable=False)
     agent_type = Column(String(20), default="user")
 
@@ -30,7 +30,7 @@ class Team(Base):
     __tablename__ = "teams"
 
     id = Column(Integer, primary_key=True, index=True)
-    team_id = Column(Integer, unique=True, nullable=False, index=True)
+    team_id = Column(BigInteger, unique=True, nullable=False, index=True)
     name = Column(String(100), nullable=False)
 
     def __repr__(self):
